addressed all comments

+++ v1/src/main/java/com/google/cloud/teleport/spanner/ddl/Column.java
@@ -126,7 +126,7 @@ public abstract class Column implements Serializable {
 
   public String typeString() {
     if (arrayLength() != null) {
-      return SizedType.typeString(type(), size(), arrayLength());
+      return SizedType.typeString(type(), size(), arrayLength().intValue());
     }
     return SizedType.typeString(type(), size());
   }
